requirements.txt
sentence-transformers>=3.0.0
torch>=2.0.0
qdrant-client>=1.7.0
numpy>=1.23.0
pydantic>=2.0.0
python-docx>=1.0.0
pypdf>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-multipart>=0.0.6
